Split types into separate file
