+++ cognite/extractorutils/uploader_extractor.py
@@ -17,36 +17,18 @@ A module containing a slightly more advanced base extractor class, sorting a gen
 """
 
 from dataclasses import dataclass
-from datetime import datetime
 from types import TracebackType
-from typing import Callable, Iterable, List, Optional, Tuple, Type, TypeVar, Union
+from typing import Callable, Iterable, Optional, Type, TypeVar
 
 from more_itertools import peekable
 
 from cognite.client import CogniteClient
-from cognite.client.data_classes import Event as _Event
-from cognite.client.data_classes import Row as _Row
 from cognite.extractorutils.base import Extractor
 from cognite.extractorutils.configtools import BaseConfig
 from cognite.extractorutils.metrics import BaseMetrics
 from cognite.extractorutils.statestore import AbstractStateStore
 from cognite.extractorutils.uploader import EventUploadQueue, RawUploadQueue, TimeSeriesUploadQueue
-
-try:
-    from typing import TypeAlias  # type: ignore
-except ImportError:
-    # Backport for python < 3.10
-    from typing_extensions import TypeAlias
-
-
-class RawRow:
-    def __init__(self, db_name: str, table_name: str, row: Union[_Row, Iterable[_Row]]):
-        self.db_name = db_name
-        self.table_name = table_name
-        if isinstance(row, Iterable):
-            self.rows = row
-        else:
-            self.rows = [row]
+from cognite.extractorutils.uploader_types import CdfTypes, Event, InsertDatapoints, RawRow
 
 
 @dataclass
@@ -62,27 +44,7 @@ class UploaderExtractorConfig(BaseConfig):
     queues: QueueConfigClass
 
 
-TimeStamp = Union[int, datetime]
-
 UploaderExtractorConfigClass = TypeVar("UploaderExtractorConfigClass", bound=UploaderExtractorConfig)
-
-
-class InsertDatapoints:
-    def __init__(
-        self,
-        *,
-        id: Optional[int] = None,
-        external_id: Optional[str] = None,
-        datapoints: Union[List[Tuple[TimeStamp, float]], List[Tuple[TimeStamp, str]]],
-    ):
-        self.id = id
-        self.external_id = external_id
-        self.datapoints = datapoints
-
-
-Event: TypeAlias = _Event
-
-CdfTypes = Union[Event, Iterable[Event], RawRow, Iterable[RawRow], InsertDatapoints, Iterable[InsertDatapoints]]
 
 
 class UploaderExtractor(Extractor[UploaderExtractorConfigClass]):
